+++ src/main/java/ticketaka/mtvs3_final_backend/member/command/domain/repository/MemberRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import ticketaka.mtvs3_final_backend.member.command.domain.model.Member;
-import ticketaka.mtvs3_final_backend.memberseat.command.domain.model.MemberSeatStatus;
+import ticketaka.mtvs3_final_backend.ticketing.memberseat.command.domain.model.MemberSeatStatus;
 
 import java.util.List;
 import java.util.Optional;
