+++ src/main/java/ticketaka/mtvs3_final_backend/ticketing/memberseat/query/repository/MemberSeatQueryRepository.java
@@ -1,9 +1,9 @@
-package ticketaka.mtvs3_final_backend.memberseat.query.repository;
+package ticketaka.mtvs3_final_backend.ticketing.memberseat.query.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ticketaka.mtvs3_final_backend.memberseat.command.domain.model.MemberSeat;
-import ticketaka.mtvs3_final_backend.memberseat.command.domain.model.MemberSeatStatus;
+import ticketaka.mtvs3_final_backend.ticketing.memberseat.command.domain.model.MemberSeat;
+import ticketaka.mtvs3_final_backend.ticketing.memberseat.command.domain.model.MemberSeatStatus;
 
 @Repository
 public interface MemberSeatQueryRepository extends JpaRepository<MemberSeat, Long> {
