+++ src/main/java/ticketaka/mtvs3_final_backend/ticketing/seat/command/application/service/SeatService.java
@@ -21,8 +21,8 @@ import ticketaka.mtvs3_final_backend.redis.drawing.domain.DrawResult;
 import ticketaka.mtvs3_final_backend.redis.drawing.domain.PaymentStatus;
 import ticketaka.mtvs3_final_backend.redis.drawing.repository.DrawResultRedisRepository;
 import ticketaka.mtvs3_final_backend.ticketing.seat.command.application.dto.SeatResponseDTO;
-import ticketaka.mtvs3_final_backend.memberseat.command.domain.model.MemberSeat;
-import ticketaka.mtvs3_final_backend.memberseat.command.domain.model.MemberSeatStatus;
+import ticketaka.mtvs3_final_backend.ticketing.memberseat.command.domain.model.MemberSeat;
+import ticketaka.mtvs3_final_backend.ticketing.memberseat.command.domain.model.MemberSeatStatus;
 import ticketaka.mtvs3_final_backend.ticketing.seat.command.domain.model.Seat;
 import ticketaka.mtvs3_final_backend.ticketing.seat.command.domain.model.SeatStatus;
 import ticketaka.mtvs3_final_backend.ticketing.seat.command.domain.repository.MemberSeatRepository;
