+++ src/main/java/ticketaka/mtvs3_final_backend/ticketing/seat/command/domain/repository/MemberSeatRepository.java
@@ -2,8 +2,8 @@ package ticketaka.mtvs3_final_backend.ticketing.seat.command.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ticketaka.mtvs3_final_backend.memberseat.command.domain.model.MemberSeat;
-import ticketaka.mtvs3_final_backend.memberseat.command.domain.model.MemberSeatStatus;
+import ticketaka.mtvs3_final_backend.ticketing.memberseat.command.domain.model.MemberSeat;
+import ticketaka.mtvs3_final_backend.ticketing.memberseat.command.domain.model.MemberSeatStatus;
 
 import java.util.Optional;
